Fix read value of CSR mip.
The read value of mip should only depend on the two bits, VSEIP and VSTIP of hvip.

This PR fix the read value of CSR mip when Sscofpmf extension enabled.

After implementing the Sscofpmf extension, hvip.LCOFIP became writable.
This caused an issue where the entire value of hvip was OR-ed when reading mip.

I revise the mip read logic to make it depend only on the hvip.VSEIP and hvip.VSTIP bits.

Co-authored-by: Zhaoyang You <zyyou1998@gmail.com>