Don't hang when invoked as background process
It's (normally) only reasonable to launch Spike as a background process
when it's used in a mode that does not need stdin.  But we should
support those use cases without hanging.  Fix the hang by catching
SIGTTOU.

For obvious reasons, `spike -d` will still hang waiting for input
until it's in the foreground.

Fixes #2133