fix infinite loop due to address overflow in pmp checking routine
Before this patch, the code checked addresses as follows: the input
address was aligned to the PMP granule boundary, and then it iterated
with granule-based increments over the referenced memory chunks. The
problem was that when the memory reference corresponded to the last
possible granule, the iteration process never ended due to a uint64_t
overflow to zero during the address increment.

The patch fixes this problem by changing the iteration from being based
on address space to being based on the indexes of the affected granules.
This approach also allows us to handle the case where the access wraps
around zero (this patch addresses only RV64 case, though)