Fix triggers for accesses wider than XLEN
I believe that @en-sc's comment here is correct:
https://github.com/riscv-software-src/riscv-isa-sim/pull/2161#discussion_r2564958203

Nevertheless, failing an assertion when someone sets a trigger on memory
accessed by a wide access is not reasonable behavior for Spike.  Better to
do something that follows the principle of least surprise, despite the
debug spec's lack of clarity on this point.